singapore_realestate_project/requirements.txt
pandas==2.0.3
numpy==1.24.3
scikit-learn==1.3.0
matplotlib==3.7.2
seaborn==0.12.2
plotly==5.16.1
folium==0.14.0
geopandas==0.13.2
selenium==4.12.0
beautifulsoup4==4.12.2
requests==2.31.0
python-dotenv==1.0.0
googlemaps==4.10.0
jupyter==1.0.0
ipython==8.15.0
